support nextjs client components
