+++ lib/MakerEnhance.d.ts
@@ -4,10 +4,5 @@ interface MakerEnhanceProps {
     index?: number;
     loadingHeight?: string;
 }
-declare global {
-    interface Window {
-        MakerEmbeds: any;
-    }
-}
 export default function MakerEnhance({ user, index, loadingHeight }: MakerEnhanceProps): JSX.Element;
 export {};
