+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "maker-enhance",
-  "version": "1.2.2",
+  "version": "2.0.0",
   "description": "React component for Maker Enhance",
   "keywords": [
     "react",
